--- apps/meteor/client/providers/VideoConfPopupProvider.tsx
@@ -1,83 +1,0 @@
-import { IRoom } from '@rocket.chat/core-typings';
-import { useCustomSound } from '@rocket.chat/ui-contexts';
-import { VideoConfPopupBackdrop } from '@rocket.chat/ui-video-conf';
-import React, { ReactElement, useState, ReactNode, useEffect, useMemo } from 'react';
-
-import { VideoConfPopupContext, VideoConfIncomingCall, VideoConfPopupPayload } from '../contexts/VideoConfPopupContext';
-import { VideoConfManager, useVideoConfIncomingCalls, useIsRinging } from '../lib/VideoConfManager';
-import VideoConfPopupPortal from '../portals/VideoConfPopupPortal';
-import VideoConfPopup from '../views/room/contextualBar/VideoConference/VideoConfPopup';
-
-const VideoConfContextProvider = ({ children }: { children: ReactNode }): ReactElement => {
-	const [outgoing, setOutgoing] = useState<VideoConfPopupPayload | undefined>();
-	const incomingCalls = useVideoConfIncomingCalls();
-	const customSound = useCustomSound();
-	const isRinging = useIsRinging();
-
-	const contextValue = useMemo(
-		() => ({
-			dispatch: (option: Omit<VideoConfPopupPayload, 'id'>): void => setOutgoing({ ...option, id: option.rid }),
-			dismiss: (): void => setOutgoing(undefined),
-			startCall: (rid: IRoom['_id']): Promise<void> => VideoConfManager.startCall(rid),
-			acceptCall: (callId: string): void => VideoConfManager.acceptIncomingCall(callId),
-			joinCall: (callId: string): Promise<void> => VideoConfManager.joinCall(callId),
-			dismissCall: (callId: string): void => {
-				VideoConfManager.dismissIncomingCall(callId);
-			},
-			rejectIncomingCall: (callId: string): void => VideoConfManager.rejectIncomingCall(callId),
-			abortCall: (): void => VideoConfManager.abortCall(),
-			useIncomingCalls: (): VideoConfIncomingCall[] => useVideoConfIncomingCalls(),
-			setPreferences: (prefs: Partial<typeof VideoConfManager['preferences']>): void => VideoConfManager.setPreferences(prefs),
-			changePreference: (key: 'cam' | 'mic', value: boolean): void => VideoConfManager.changePreference(key, value),
-			useIsRinging: (): boolean => useIsRinging(),
-		}),
-		[],
-	);
-
-	const popups = useMemo(
-		() =>
-			incomingCalls
-				.filter((incomingCall) => !incomingCall.dismissed)
-				.map((incomingCall) => ({ id: incomingCall.callId, rid: incomingCall.rid, isReceiving: true })),
-		[incomingCalls],
-	);
-
-	useEffect(() => {
-		if (isRinging) {
-			customSound.play('calling');
-			const soundInterval = setInterval(() => {
-				customSound.play('calling');
-			}, 3000);
-
-			return (): void => {
-				customSound.pause('calling');
-				clearInterval(soundInterval);
-			};
-		}
-	}, [customSound, isRinging]);
-
-	return (
-		<VideoConfPopupContext.Provider value={contextValue}>
-			{children}
-			{(outgoing || popups?.length > 0) && (
-				<VideoConfPopupPortal>
-					<VideoConfPopupBackdrop>
-						{(outgoing ? [outgoing, ...popups] : popups).map(({ id, rid, isReceiving }, index) => (
-							<VideoConfPopup
-								key={id}
-								id={id}
-								rid={rid}
-								isReceiving={isReceiving}
-								position={(index + 1) * 10}
-								current={index}
-								total={popups.length}
-							/>
-						))}
-					</VideoConfPopupBackdrop>
-				</VideoConfPopupPortal>
-			)}
-		</VideoConfPopupContext.Provider>
-	);
-};
-
-export default VideoConfContextProvider;
