+++ apps/meteor/client/views/room/contextualBar/VideoConference/VideoConfPopups/VideoConfPopup/ReceivingPopup.tsx
@@ -11,42 +11,46 @@ import {
 	VideoConfButton,
 	VideoConfPopupFooter,
 	VideoConfPopupTitle,
+	VideoConfPopupIndicators,
+	VideoConfPopupClose,
 } from '@rocket.chat/ui-video-conf';
 import React, { ReactElement } from 'react';
 
-import ReactiveUserStatus from '../../../../../components/UserStatus/ReactiveUserStatus';
-import RoomAvatar from '../../../../../components/avatar/RoomAvatar';
-import { useChangePreference } from '../../../../../contexts/VideoConfPopupContext';
+import ReactiveUserStatus from '../../../../../../components/UserStatus/ReactiveUserStatus';
+import RoomAvatar from '../../../../../../components/avatar/RoomAvatar';
+import { useSetPreferences } from '../../../../../../contexts/VideoConfContext';
 
-type CallingPopupProps = {
+type ReceivingPopupProps = {
 	id: string;
 	room: IRoom;
+	position: number;
+	current: number;
+	total: number;
 	onClose: (id: string) => void;
+	onMute: (id: string) => void;
+	onConfirm: () => void;
 };
 
-const CallingPopup = ({ room, onClose, id }: CallingPopupProps): ReactElement => {
+const ReceivingPopup = ({ id, room, position, current, total, onClose, onMute, onConfirm }: ReceivingPopupProps): ReactElement => {
 	const t = useTranslation();
 	const userId = useUserId();
 	const directUserId = room.uids?.filter((uid) => uid !== userId).shift();
 	const { controllersConfig, handleToggleMic, handleToggleCam } = useVideoConfControllers();
-	const changePreference = useChangePreference();
+	const setPreferences = useSetPreferences();
 
-	const handleToggleMicPref = useMutableCallback(() => {
-		changePreference('mic', !controllersConfig.mic);
-		handleToggleMic();
-	});
-
-	const handleToggleCamPref = useMutableCallback(() => {
-		changePreference('cam', !controllersConfig.cam);
-		handleToggleCam();
+	const handleJoinCall = useMutableCallback(() => {
+		setPreferences(controllersConfig);
+		onConfirm();
 	});
 
 	return (
-		<VideoConfPopup>
+		<VideoConfPopup position={position}>
 			<VideoConfPopupContent>
 				{/* Design Team has planned x48 */}
+				<VideoConfPopupClose title={t('Close')} onClick={(): void => onMute(id)} />
 				<RoomAvatar room={room} size='x40' />
-				<VideoConfPopupTitle text='Calling' icon='phone-out' counter />
+				{current && total ? <VideoConfPopupIndicators current={current} total={total} /> : null}
+				<VideoConfPopupTitle text='Incoming call from' icon='phone-in' />
 				{directUserId && (
 					<Box display='flex' alignItems='center' mbs='x8'>
 						<ReactiveUserStatus uid={directUserId} />
@@ -64,20 +68,23 @@ const CallingPopup = ({ room, onClose, id }: CallingPopupProps): ReactElement =>
 						text={controllersConfig.mic ? t('Mic_on') : t('Mic_off')}
 						title={controllersConfig.mic ? t('Mic_on') : t('Mic_off')}
 						icon={controllersConfig.mic ? 'mic' : 'mic-off'}
-						onClick={handleToggleMicPref}
+						onClick={handleToggleMic}
 					/>
 					<VideoConfController
 						primary={controllersConfig.cam}
 						text={controllersConfig.cam ? t('Cam_on') : t('Cam_off')}
 						title={controllersConfig.cam ? t('Cam_on') : t('Cam_off')}
 						icon={controllersConfig.cam ? 'video' : 'video-off'}
-						onClick={handleToggleCamPref}
+						onClick={handleToggleCam}
 					/>
 				</VideoConfPopupControllers>
 				<VideoConfPopupFooter>
+					<VideoConfButton primary onClick={handleJoinCall}>
+						{t('Accept')}
+					</VideoConfButton>
 					{onClose && (
-						<VideoConfButton primary icon='phone-disabled' onClick={(): void => onClose(id)}>
-							{t('Cancel')}
+						<VideoConfButton danger onClick={(): void => onClose(id)}>
+							{t('Decline')}
 						</VideoConfButton>
 					)}
 				</VideoConfPopupFooter>
@@ -86,4 +93,4 @@ const CallingPopup = ({ room, onClose, id }: CallingPopupProps): ReactElement =>
 	);
 };
 
-export default CallingPopup;
+export default ReceivingPopup;
