+++ apps/meteor/client/views/room/contextualBar/VideoConference/VideoConfPopups/index.ts
@@ -1,0 +1,1 @@
+export { default } from './VideoConfPopups';
